requirements.txt
Flask==1.1.2
beautifulsoup4
requests
google-cloud-datastore
google-cloud-storage
